feat: migrate availability api to firebase
